Exercise 4 - part 1 (#6)
* add product form working

* category form adding added

* added basket add form

* added basketitem form adding

* cleaned up

* added payment form adding

* added order form adding

* added repository for promo

* added promo adding form

* prepared favoriteItemsControllers to forms

* added favoriteItem adding form

* prepared product question controller to forms

* added product Question adding form

* prepared production decsription to use forms

* added product description adding form

+++ exercise-2/app/models/Promo.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Promo(var id: Float, var productId: Float, var looseAmount: Int)
+case class Promo(var id: Long, var productId: Long, var looseAmount: Int)
 
 object Promo{
   implicit val promoFormat: OFormat[Promo] = Json.format[Promo];
